[5/N][Refactor] torchair model runner refactor
